Refactor the CLI into two classes, one for command line logic, and another for invoking the the SubmissionEngine.  This allows the business logic for invoking the SubmissionEngine to be tested without having System.exit(...) being invoked.
