+++ nihms-cli/pom.xml
@@ -45,7 +45,7 @@
                             <transformers>
                                 <transformer implementation=
                                                      "org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>org.dataconservancy.nihms.cli.NihmsSubmissionApp</mainClass>
+                                    <mainClass>org.dataconservancy.nihms.cli.NihmsSubmissionCLI</mainClass>
                                 </transformer>
                             </transformers>
                         </configuration>
